+++ src/dime/util.clj
@@ -10,6 +10,9 @@
 (ns dime.util)
 
 
+;; ----- metadata attribute keys -----
+
+
 (def ^:dynamic *expose-meta-key* :expose)
 
 
@@ -22,19 +25,55 @@
 (def ^:dynamic *post-inject-meta-key* :post-inject)
 
 
+;; ----- injection utility fns -----
+
+
 (defn pre-inject-default
-  "Arity-2 fn that may be used as a pre-inject fn for default behavior, that is to return the injectable intact."
+  "Identity pre injector - return the injectable object intact."
   [injectable deps]
   injectable)
 
 
 (defn post-inject-default
-  "Arity-3 fn that may be used as a post-inject fn for default behavior, that is to return the injectable intact."
+  "Identity post injector - return the injected object intact."
   [injected node-id deps]
   injected)
 
 
 (defn post-inject-invoke
-  "Arity-3 fn that may be used as a post-inject fn to invoke the injected as if it were a no-argument fn."
+  "Invoke the injected as if it were a no-argument fn."
   [injected node-id deps]
   (injected))
+
+
+(defn post-inject-alter
+  "Bind specified var to the injected."
+  [the-var]
+  (fn [injected node-id deps]
+    (alter-var-root the-var (constantly injected))))
+
+
+(defn comp-pre-inject
+  "Compose multiple pre injectors into one. Like clojure.core/comp, except it accepts arity-2 fns as pre-injectors."
+  [& pre-injectors]
+  (if (seq pre-injectors)
+    (fn [injectable deps]
+      (as-> pre-injectors $
+        (map (fn [each-pre-injector deps]
+               #(each-pre-injector % deps)) $ (repeat deps))
+        (apply comp $)
+        ($ injectable)))
+    pre-inject-default))
+
+
+(defn comp-post-inject
+  "Compose multiple post injectors into one. Like clojure.core/comp, except it accepts arity-3 fns as post-injectors."
+  [& post-injectors]
+  (if (seq post-injectors)
+    (fn [injected node-id deps]
+      (as-> post-injectors $
+        (map (fn [each-post-injector node-id deps]
+               #(each-post-injector % node-id deps)) $ (repeat node-id) (repeat deps))
+        (apply comp $)
+        ($ injected)))
+    post-inject-default))